DSM]-PEPPER-91-improve-shared-learnings-codebase
